[Refectoring]LandingPage 컴포넌트화하여 분리
